requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
nats-py>=2.3.0
pyjwt>=2.6.0
python-dotenv>=1.0.0
pydantic>=1.10.7
httpx>=0.24.0
boto3>=1.26.0
cryptography>=40.0.0
python-multipart>=0.0.6
ulid-py>=1.1.0
requests>=2.28.0
pydantic-settings>=2.0.0
email-validator>=2.0.0